+++ common/src/main/java/com/faboslav/friendsandfoes/common/mixin/RaidMemberMixin.java
@@ -1,16 +1,18 @@
 package com.faboslav.friendsandfoes.common.mixin;
 
 import com.faboslav.friendsandfoes.common.FriendsAndFoes;
+import com.llamalad7.mixinextras.injector.wrapoperation.Operation;
+import com.llamalad7.mixinextras.injector.wrapoperation.WrapOperation;
+import net.minecraft.world.entity.EntityType;
 import net.minecraft.world.entity.raid.Raid;
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.injection.At;
-import org.spongepowered.asm.mixin.injection.ModifyArg;
 import org.spongepowered.asm.mixin.injection.Slice;
 
 @Mixin(Raid.RaiderType.class)
 public final class RaidMemberMixin
 {
-	@ModifyArg(
+	@WrapOperation(
 		method = "<clinit>",
 		slice = @Slice(
 			from = @At(
@@ -20,12 +22,15 @@ public final class RaidMemberMixin
 		),
 		at = @At(
 			value = "INVOKE",
-			target = "Lnet/minecraft/world/entity/raid/Raid$RaiderType;<init>(Ljava/lang/String;ILnet/minecraft/world/entity/EntityType;[I)V",
-			ordinal = 0
+			target = "Lnet/minecraft/world/entity/raid/Raid$RaiderType;<init>(Ljava/lang/String;ILnet/minecraft/world/entity/EntityType;[I)V"
 		)
 	)
-	private static int[] friendsandfoes_updateCountInWave(
-		int[] countInWave
+	private static void friendsandfoes$wrapEvokerCtor(
+		String name,
+		int id,
+		EntityType<?> type,
+		int[] countInWave,
+		Operation<Void> original
 	) {
 		if (
 			(
@@ -37,9 +42,9 @@ public final class RaidMemberMixin
 				&& FriendsAndFoes.getConfig().enableIceologerInRaids
 			)
 		) {
-			return new int[]{0, 0, 0, 0, 0, 1, 1, 1};
+			original.call(name, id, type, new int[]{0, 0, 0, 0, 0, 1, 1, 1});
+		} else {
+			original.call(name, id, type, countInWave);
 		}
-
-		return new int[]{0, 0, 0, 0, 0, 1, 1, 2};
 	}
 }