+++ common/src/main/java/com/faboslav/friendsandfoes/common/mixin/RaidMixin.java
@@ -3,13 +3,12 @@ package com.faboslav.friendsandfoes.common.mixin;
 import com.faboslav.friendsandfoes.common.FriendsAndFoes;
 import com.faboslav.friendsandfoes.common.entity.IceologerEntity;
 import com.faboslav.friendsandfoes.common.entity.IllusionerEntity;
+import com.llamalad7.mixinextras.injector.wrapmethod.WrapMethod;
+import com.llamalad7.mixinextras.injector.wrapoperation.Operation;
 import net.minecraft.core.BlockPos;
 import net.minecraft.world.entity.raid.Raid;
 import net.minecraft.world.entity.raid.Raider;
 import org.spongepowered.asm.mixin.Mixin;
-import org.spongepowered.asm.mixin.injection.At;
-import org.spongepowered.asm.mixin.injection.Inject;
-import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
 
 //? if >= 1.21.5 {
 import net.minecraft.server.level.ServerLevel;
@@ -18,10 +17,8 @@ import net.minecraft.server.level.ServerLevel;
 @Mixin(Raid.class)
 public final class RaidMixin
 {
-	@Inject(
-		method = "joinRaid",
-		at = @At("HEAD"),
-		cancellable = true
+	@WrapMethod(
+		method = "joinRaid"
 	)
 	public void friendsandfoes_addRaider(
 		//? if >= 1.21.5 {
@@ -30,8 +27,8 @@ public final class RaidMixin
 		int wave,
 		Raider raider,
 		BlockPos pos,
-		boolean existing,
-		CallbackInfo ci
+		boolean isRecruited,
+		Operation<Void> original
 	) {
 		if (
 			(
@@ -48,7 +45,17 @@ public final class RaidMixin
 				)
 			)
 		) {
-			ci.cancel();
+			return;
 		}
+
+		original.call(
+			//? if >= 1.21.5 {
+			serverLevel,
+			//?}
+			wave,
+			raider,
+			pos,
+			isRecruited
+		);
 	}
 }