+++ common/src/main/java/com/faboslav/friendsandfoes/common/mixin/VillagerHostilesSensorMixin.java
@@ -2,51 +2,48 @@ package com.faboslav.friendsandfoes.common.mixin;
 
 import com.faboslav.friendsandfoes.common.init.FriendsAndFoesEntityTypes;
 import com.google.common.collect.ImmutableMap;
+import com.llamalad7.mixinextras.injector.wrapmethod.WrapMethod;
+import com.llamalad7.mixinextras.injector.wrapoperation.Operation;
 import net.minecraft.world.entity.EntityType;
 import net.minecraft.world.entity.LivingEntity;
 import net.minecraft.world.entity.ai.sensing.VillagerHostilesSensor;
 import org.spongepowered.asm.mixin.Mixin;
-import org.spongepowered.asm.mixin.injection.At;
-import org.spongepowered.asm.mixin.injection.Inject;
-import org.spongepowered.asm.mixin.injection.callback.CallbackInfoReturnable;
 
 @Mixin(VillagerHostilesSensor.class)
 public final class VillagerHostilesSensorMixin
 {
-	private static final ImmutableMap<EntityType<?>, Float> CUSTOM_SQUARED_DISTANCES_FOR_DANGER;
+	private static final ImmutableMap<EntityType<?>, Float> CUSTOM_SQUARED_DISTANCES_FOR_DANGER = ImmutableMap.<EntityType<?>, Float>builder().put(FriendsAndFoesEntityTypes.ILLUSIONER.get(), 12.0F).put(FriendsAndFoesEntityTypes.ICEOLOGER.get(), 12.0F).build();
 
-	static {
-		CUSTOM_SQUARED_DISTANCES_FOR_DANGER = ImmutableMap.<EntityType<?>, Float>builder().put(FriendsAndFoesEntityTypes.ICEOLOGER.get(), 12.0F).build();
-	}
-
-	@Inject(
-		at = @At("HEAD"),
-		method = "isClose",
-		cancellable = true
+	@WrapMethod(
+		method = "isClose"
 	)
-	private void friendsandfoes_isCloseEnoughForDanger(
-		LivingEntity villager,
+	private boolean friendsandfoes_isCloseEnoughForDanger(
+		LivingEntity attacker,
 		LivingEntity target,
-		CallbackInfoReturnable<Boolean> cir
+		Operation<Boolean> original
 	) {
+		if(original.call(attacker, target)) {
+			return true;
+		}
+
 		if (CUSTOM_SQUARED_DISTANCES_FOR_DANGER.containsKey(target.getType())) {
 			float distance = CUSTOM_SQUARED_DISTANCES_FOR_DANGER.get(target.getType());
-			boolean isCloseEnoughForDanger = target.distanceToSqr(villager) <= (double) (distance * distance);
-			cir.setReturnValue(isCloseEnoughForDanger);
+			return target.distanceToSqr(attacker) <= (double) (distance * distance);
 		}
+
+		return false;
 	}
 
-	@Inject(
-		at = @At("HEAD"),
-		method = "isHostile",
-		cancellable = true
+	@WrapMethod(
+		method = "isHostile"
 	)
-	private void friendsandfoes_isHostile(
-		LivingEntity entity,
-		CallbackInfoReturnable<Boolean> cir
+	private boolean friendsandfoes$isHostile(
+		LivingEntity entity, Operation<Boolean> original
 	) {
-		if (CUSTOM_SQUARED_DISTANCES_FOR_DANGER.containsKey(entity.getType())) {
-			cir.setReturnValue(true);
+		if(original.call(entity)) {
+			return true;
 		}
+
+		return CUSTOM_SQUARED_DISTANCES_FOR_DANGER.containsKey(entity.getType());
 	}
 }